Added 'switches' management command to create / delete switches listed in HASHBROWN_SWITCH_DEFAULTS.
Running 'python manage.py switches' will create any switches listed in the settings which are
not in the database already. Using '--delete' will also delete any switches which are in the
database but not in the settings, prompting for confirmation first (unless you also use
'--force' which will skip the confirmation and go straight to deleting).