Shorten the max_length of label to 190 characters
`label` requires a unique index. On MySQL the max length of a unique indexed column is 767 bytes.

When using the utf8 collation this works because utf8 isn't *real* utf8, it only encodes characters between 1-3 bytes. `PILE OF POO` emojis use 4.

To correctly store those emojis you can use true utf8 on MySQL using the utf8m4 encoding, however, this means that only 191 characters will fit in 767 bytes and on that encoding the CREATE TABLE statement explodes if the field is longer than that.

So the fix is to reduce the length from 200, to a nice round 190 which is below the 191 limit.